requirements.txt
app-version==1.0.1
asgiref==3.2.10
astroid==2.4.2
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.8
django-appconf==1.0.4
django-cors-headers==3.4.0
django-heroku==0.3.1
django-permission==1.0.5
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
django-session-timeout==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.12.6
Pillow==7.2.0
psycopg2==2.8.5
Pygments==2.6.1
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
